style: rename "Layout" to "layout", add a barrel
This is a category of components (like /common),
than a single component.
