+++ services/orchest-webserver/client/src/jobs-view/JobsView.tsx
@@ -1,4 +1,4 @@
-import { LayoutWithSidePanel } from "@/components/Layout/LayoutWithSidePanel";
+import { LayoutWithSidePanel } from "@/components/layout";
 import { useSendAnalyticEvent } from "@/hooks/useSendAnalyticEvent";
 import { siteMap } from "@/routingConfig";
 import React from "react";
