+++ services/orchest-webserver/client/src/settings-view/SettingsView.tsx
@@ -1,6 +1,6 @@
 import { Code } from "@/components/common/Code";
 import { PageTitle } from "@/components/common/PageTitle";
-import { Layout } from "@/components/Layout";
+import { Layout } from "@/components/layout";
 import { useAppContext } from "@/contexts/AppContext";
 import { useGlobalContext } from "@/contexts/GlobalContext";
 import { useCustomRoute } from "@/hooks/useCustomRoute";
